requirements_model.txt
--extra-index-url https://download.pytorch.org/whl/cu118

torch>=2.3.1
transformers==4.45.2
tokenizers==0.20.3
accelerate==0.34.2
safetensors>=0.4.3
numpy==1.26.4
autoawq==0.2.6

fastapi
uvicorn
hf_transfer
runpod